+++ components/ledger/ledger-utxo-flow/src/main/java/net/corda/ledger/utxo/flow/impl/groupparameters/verifier/package-info.java
@@ -1,0 +1,4 @@
+@Export
+package net.corda.ledger.utxo.flow.impl.groupparameters.verifier;
+
+import org.osgi.annotation.bundle.Export;
